TS => MJS: types/sorted_list
Migrate fjs/types/sorted_list/module.f.ts to module.f.mjs under the
stage-1 TypeScript-to-mjs migration, folded into the types/patricia_trie
PR since both are independent dependency leaves. All of its runtime
dependencies (function/compare, list, nullable, function) are already
.f.mjs.

proof.f.ts stays TypeScript for now: it still imports the unmigrated
media/json and object modules, which the migration allows (only the
import specifier is updated).

The curried, doubly-generic helpers (genericMerge, merge, cmpReduce,
intersectReduce, find) use the per-arrow @template/@param JSDoc style
already established in fjs/types/array/module.f.mjs's isTuple, rather
than a single @type {<T, S>(...) => ...} on the whole chain — nested
generic-function-value composition otherwise loses inference and widens
to unknown.

Update the two remaining .f.ts importers (range_map/module.f.ts,
sorted_set/module.f.ts) and re-anchor the two todo docs that quoted the
old .f.ts source and line numbers.
